Change global variable handling
Now every reference to non local variable is a reference to global
variable. Name of each defined/accessed/assigned global variable
must be added to the list of globals exactly once.

All definitions inside local scope are local, all definitions outside
local scope are global.

Definitions inside objects denote members and there can't be a single
member name defined twice (this is checked in parser for simplicity).

So even though globals and objects ultimately compile to Classes, they
behave differently with regards to duplicate definitions. This is so
that the behavior is in line with the AST interpreter.